Refactor jdbc module code
