+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/actions/BatchUpdate.java
@@ -25,7 +25,7 @@ import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinax.jdbc.Constants;
-import org.ballerinax.jdbc.SQLDatasource;
+import org.ballerinax.jdbc.datasource.SQLDatasource;
 import org.ballerinax.jdbc.statement.BatchUpdateStatement;
 import org.ballerinax.jdbc.statement.SQLStatement;
