+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/functions/CreateClient.java
@@ -24,7 +24,7 @@ import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinax.jdbc.Constants;
-import org.ballerinax.jdbc.SQLDatasourceUtils;
+import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
 
 import java.util.UUID;
