+++ stdlib/jdbc/src/test/resources/test-src/sql/sql_actions_negative_test.bal
@@ -133,7 +133,7 @@ function testGeneratedKeyOnInsertError() returns @tainted [string, boolean, bool
     return [reason, isMessageExist, isSqlErrorCodeExist, isSqlStateExist];
 }
 
-function testUpdateReslt() returns int|string {
+function testUpdateResult() returns int|string {
     jdbc:Client testDB = new({
             url: "jdbc:h2:file:./target/tempdb/TEST_SQL_CONNECTOR_H2",
             username: "SA",
